==> analysis/15_d5_true_and_false/scripts/03_KD_comparison.Rmd <==
---
title: "KD comparison"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(tidyverse)
library(kableExtra)
library(ggpubr)
```

# Compare with other KD IRfinder results on the paper

## Data on the paper

```{r out.width=200}
knitr::include_graphics('../figs/paper_KD_results.png')
```

```{r}
# change the above data into a tibble
dat <- tibble(
  gene = c('TIA1','SRSF1','U2AF2','PCBP1','PCBP2','PTBP1','SRSF7','FUS'),
  decreased = c(52, 214, 141, 53, 43, 90, 46, 221),
  increased = c(1849, 2088, 3204, 2775, 1886, 2122, 1831, 227)
)
```

## Our data: nominal P value < 0.05 events and differentiate them to be increased and decreased


```{r}
# filter to have the nominal P value < 0.05 events to be significant 
# then put them to be IR increased and decreased by comparing A_IRratio with B_IRratio
data_path <- '../../05_IRfinder/data/rawIRfinderResult/'
fns <- list.files(data_path, pattern = 'txt', full.names = T)

sig <- fns %>%
  map(~ read_tsv(.,comment = '#',col_names = T)  %>%
        mutate(Start = Start + 1) %>%
        mutate(locus = paste(Chr,paste(Start, End, sep = '-'), sep = ':' ) ) %>%
        rename_all(funs(stringr::str_replace_all(., '-', '_') )) %>%
        separate(names(.)[4], into = c('Symbol', 'GeneID', 'Category'), sep = '/') %>%
        mutate(GeneID = str_replace_all(string = GeneID, pattern = '\\..*$', replacement = '')) %>%
        filter(p_diff < 0.05) %>%
        mutate(direction = ifelse(A_IRratio > B_IRratio, 'increased','decreased')) %>%
        split(.$direction) %>%
        map_dfc(~ nrow(.)) 
      ) %>%
  bind_rows %>%
  mutate(group = str_match(fns, 'RF..'))
        

sig %>%
  kable(caption = 'Num of increased and decreased IR events in each comparison') %>%
  kable_styling(full_width = F, position = 'left')

```


## Get the individual of 14 and 89, also the union and intersect of the two

```{r}
# get the locus only so I can compute the intersect and union
sig_locus <- fns %>%
  map(~ read_tsv(.,comment = '#',col_names = T)  %>%
        mutate(Start = Start + 1) %>%
        mutate(locus = paste(Chr,paste(Start, End, sep = '-'), sep = ':' ) ) %>%
        rename_all(funs(stringr::str_replace_all(., '-', '_') )) %>%
        separate(names(.)[4], into = c('Symbol', 'GeneID', 'Category'), sep = '/') %>%
        mutate(GeneID = str_replace_all(string = GeneID, pattern = '\\..*$', replacement = '')) %>%
        filter(p_diff < 0.05) %>%
        mutate(direction = ifelse(A_IRratio > B_IRratio, 'increased','decreased'))%>%
        split(.$direction) 
      )

up14 <- sig_locus[[1]][['increased']]$locus
down14 <- sig_locus[[1]][['decreased']]$locus

up89 <- sig_locus[[2]][['increased']]$locus
down89 <- sig_locus[[2]][['decreased']]$locus

# get the intersect and union of those
RF14 <- list(up14 = up14, down14 = down14)
RF89 <- list(up89 = up89, down89 = down89)

overlap <- map2(RF14, RF89, intersect)  %>%
  map_dfc(length) %>%
  rename(increased = names(.)[1],
         decreased = names(.)[2])

Union <- map2(RF14, RF89, union) %>%
  map_dfc(length) %>%
  rename(increased = names(.)[1],
         decreased = names(.)[2])

# bind them together
two <- bind_rows(overlap, Union) %>%
  mutate(group = c('intersect', 'union'))

our <- bind_rows(sig, two) %>%
  dplyr::select(gene = group, decreased, increased) %>%
  mutate(gene = str_glue('SDE2 {gene}'))

our %>%
  kable() %>%
  kable_styling(full_width = F, position = 'left')
```


## Combine our data with the paper data, and divide by the total number of possible RI events to get a percentage.

```{r}
# the total number of possible IR events is 
total <- read_csv('../data/unique_introns_for_expressive_genes.csv') %>%
  filter(intronLen > 30) %>%
  nrow

# combine with the paper's dataset
final <- bind_rows(dat, our) %>%
  mutate(decreased_in_perc = paste(round(decreased/total*100, 2), '%', sep = ''),
         increased_in_perc = paste(round(increased/total*100, 2), '%', sep = ''))

final %>%
  kable(caption = 'Combined data in number and fraction') %>%
  kable_styling(full_width = F, position = 'left')
```

```{r}
# write out final so that I could plot in other scripts
final %>%
  write_csv('../output/paper and our KD events compare.csv')
```


```{r}
# make the above "increased" in percent a bar plot just to show the difference 
# remove the union row, change the siRNA's name, add a factor to tell our data and paper data apart
final %>%
  mutate(gene = str_replace(gene, 'RF', 'Si'))  %>%
  mutate(group = if_else(str_detect(gene, 'SDE2'), 'our', 'paper')) %>%
  ggbarplot(x = 'gene', y ='increased',
            color = 'group', fill = 'group',
            ylab = 'Number of events',
            title = 'Comparison of increased IR events after gene KD')  +
  ylim(0, 7900) +
  geom_text(aes(label = increased_in_perc), vjust = -1, size = 3 ) +
  scale_color_manual(values = c('#EEBF2B', '#1175C0'))  +
  scale_fill_manual(values = c('#EEBF2B', '#1175C0'))  +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
      axis.text=element_text(size=8),
      axis.title.x=element_blank(),
      legend.position = "none") +
  ggsave('../figs/KD_comparison_bar.jpg', width = 7, height = 5, dpi = 300)

  
```
