==> analysis/15_d5_true_and_false/scripts/04_Fig4.Rmd <==
---
title: "04_Fig4"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---
## This is for day 5. Intersect.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(tidyverse)
library(ggpubr)
library(scales)
library(extrafont)
library(grid)
library(cowplot)
```

```{r}
# load the result table
res <- read_csv('../../05_IRfinder/data/02_intersect_sig_and_intersect_non_sig_all_other_info.csv') 
```

# SiSDE2-1 (Si14 only)

```{r}
table(res$Significance) 
```

```{r}
# parameters about the size of the text and the axis text(for publication level of plots)
axis_text_size <- 9
general_text_size <- 10
rotation_angle <- 30
our_font <- 'Arial'
```

## a: SDE2 IR events VS IRfinder events

```{r}
# make the 14 89 named si-1 and si-2
compare <- read_csv('../output/paper and our KD events compare.csv') %>%
  mutate(gene = str_replace(gene, ' RF14', '-1'),
         gene = str_replace(gene, ' RF89', '-2')) %>% 
  mutate(gene = if_else(str_detect(gene, 'SDE2'),
                        str_glue('si{gene}'),
                        gene))

compare_plot <- compare %>%
  mutate(gene = str_replace(gene, 'RF', 'Si'))  %>%
  mutate(group = if_else(str_detect(gene, 'SDE2'), ' SDE2 data', ' ENCODE data ')) %>%
  ggbarplot(x = 'gene', y ='increased',
            color = 'group', fill = 'group',
            ylab = 'Number of events')  +
  ylim(0, 7800) +
  geom_text(aes(label = increased_in_perc), vjust = -1, size = 2 ) +
  scale_color_manual(values = c('#EEBF2B', '#1175C0'))  +
  scale_fill_manual(values = c('#EEBF2B', '#1175C0'))  +
  theme(text = element_text(family = our_font),
        axis.text.x = element_text(angle = rotation_angle, hjust = 1),
        axis.text=element_text(size=axis_text_size),
        axis.title.x=element_blank(),
        axis.title.y=element_text(size = general_text_size),
        legend.title = element_blank())
```

## Position

```{r}
pos <- res %>%
  mutate(Significance = factor(Significance, levels = c('Sig','Non-sig'))) %>% 
  gghistogram(x = 'RI_prop',
              palette = 'lancet',
              color = 'white', fill = 'Significance', 
              bins = 40,
              add = "median",
              xlab = 'RI position') + 
  scale_color_manual(values = c('#EC0000','#00468B')) +
  scale_fill_manual(values = c('#EC0000','#00468B')) +
  theme(text = element_text(family = our_font),
        axis.text=element_text(size=axis_text_size),
        axis.title=element_text(size = general_text_size),
        axis.title.x=element_blank(),
        legend.title = element_blank())
```

## c: Length 
 
```{r}
# add a third box which is all the background introns other than the first two

# change the format of the res sig table in the below way
IRfinder <- res %>% 
  dplyr::select(locus,
                intronLen = RI_len,
                category = Significance,
                GC_per,
                ss5_score,
                ss3_score)

# load the background introns
non_AS <- read_csv('../../14_Day3/analysis/05_IRfinder/data/03_non_AS_seq_info.csv') %>% 
  filter(intronLen > 30) %>% 
  filter(!locus %in% IRfinder$locus) %>% 
  dplyr::select(locus,
                intronLen ,
                category ,
                GC_per,
                ss5_score,
                ss3_score) 

all <- bind_rows(
 IRfinder,
 non_AS %>% 
    mutate(category = 'Non-AS')
) %>%
  mutate(category = factor(category, levels = c('Sig','Non-sig','Non-AS')))


# wanna add a label of how many events in each type of Significance
tally <- table(all$category) %>%
  data.frame %>%
  dplyr::rename(category = names(.)[1])

all <- all %>%
  split(.$category) %>%
  map_df(~  mutate(.data = ., Num = nrow(.)))

# annotation_height to be the min of the Non-AS
annotation_height <- all %>%
  filter(category == 'Non-AS') %>%
  summarise(mini = min(intronLen))%>%
  unlist
  
len <- all %>%
  ggboxplot(x = "category", y = 'intronLen',  color = 'category',
            ylab = 'Length of retained intron', xlab = '') +
  scale_y_log10(breaks = trans_breaks("log10", function(x) 10^x),
              labels = trans_format("log10", math_format(10^.x)))  +
  annotation_logticks(sides = "l")  +
  scale_color_manual(values = c('#EC0000','#00468B', '#374E54'))  +
  stat_compare_means(comparisons = list(c('Sig', 'Non-sig'),
                                        c('Non-sig','Non-AS'),
                                        c('Sig','Non-AS')),
                     label = "p.signif") +
  geom_text(data = tally, aes(x = category, y = annotation_height, label = Freq)) +
  theme(text = element_text(family = our_font),
        axis.text=element_text(size=axis_text_size),
        axis.title.y=element_text(size = general_text_size),
        legend.position="none")
```
 
## d: GC content

```{r}
all <- bind_rows(
 IRfinder,
 non_AS 
) %>% 
  mutate(category = factor(category, levels = c('Sig','Non-sig','Non-AS-short','Non-AS-long')))

# add the number of the events in the 4 groups
tally_4 <- table(all$category) %>%
  data.frame %>%
  dplyr::rename(category = names(.)[1])

annotation_height_GC <- all %>%
  filter(category == 'Non-AS-long') %>%
  summarise(mini = min(GC_per))%>%
  unlist
```

```{r}
gc <- all %>%
    ggboxplot(x = 'category', y = 'GC_per',
              color = 'category', 
              xlab = 'GC% ',
              ylab = 'GC%') +
    scale_y_continuous(breaks=seq(0, 100, 25)) +
    stat_compare_means(comparisons = list(c('Sig', 'Non-sig'),
                                          c('Non-AS-short','Non-AS-long'),
                                          c('Sig','Non-AS-short')), label = "p.signif")+   
    scale_color_manual(values = c( '#EC0000','#00468B','#42B440','#1799B1')) +
    geom_text(data = tally_4, aes(x = category, y = annotation_height_GC, label = Freq)) +
    theme(text = element_text(family = our_font),
          legend.position="none",
          axis.title.x=element_blank(),
          axis.text=element_text(size=axis_text_size),
          axis.title.y=element_text(size = general_text_size)) 
```

## e: Splice site strength  

```{r}
#   filter(ss5_score >= -7.2 & ss3_score >= 1.4) %>% 
list_strength <- all %>% 
  dplyr::select(locus, category, ss5_score, ss3_score) %>%
  gather(key = 'type', value = 'score', ss5_score:ss3_score) %>% 
  split(.$type)

# we are doing the splitting and combining because the ss5 and ss3 have to be filtered separately, thanks Adam for this brilliant idea!
df_strength <- bind_rows(
  list_strength$ss3_score %>% 
    filter(score >= 1.4),
  list_strength$ss5_score %>% 
    filter(score >= -7.2)
)

table(df_strength$type)
  
strength <- df_strength %>% 
  mutate(type = str_replace(type, '_score','')) %>% 
  mutate(type = factor(type, levels = c('ss5','ss3'))) %>% 
  ggboxplot(x = 'category', y = 'score',
            ylab = 'Splice site score', xlab = '', color = 'category')   +
  stat_compare_means(comparisons = list(c('Sig', 'Non-sig'),
                                        c('Non-AS-short','Non-AS-long'),
                                        c('Sig','Non-AS-short')), 
                     label = "p.signif",
                     label.y = c(17, 18, 21)) +   
  scale_color_manual(values = c( '#EC0000','#00468B','#42B440','#1799B1')) +
  facet_grid(~type) +
  theme(text = element_text(family = our_font),  
        axis.text=element_text(size=axis_text_size),
        axis.text.x = element_text(angle = rotation_angle, hjust = 1),
        axis.title.y=element_text(size = general_text_size),
        panel.border = element_rect( colour = "black",fill=NA, size=0.8),
        axis.line = element_line(colour = "black", size = 0),
        strip.text = element_text(size=general_text_size),
        legend.position="none") +
  coord_cartesian(ylim = c(0, 22))      
  
```
 
## GSEA 

```{r}
# the gsea of the significant genes
res %>%
  filter(Significance == 'Sig') %>%
  distinct(ensembl_gene_id)  %>%
  write_csv('../data/unique_gene_id_for_significant_day5_intersection.csv', col_names = F)

# plot the -log10(q) value from Metascape results (top 10)
gsea <- read_csv('../output/Metascape/day5_sig_gene/all/Enrichment_GO/GO_AllLists.csv') %>%
  top_n(n = -10, wt = `Log(q-value)`) %>%
  mutate(neg_log_q = abs(`Log(q-value)`)) %>%
  mutate(Description = str_replace(Description, ' \\(..+$','')) %>% 
    mutate(Description = if_else(str_detect(Description, 'ribonucleoprotein complex biogenesis'), 'ribonucleoprotein\ncomplex biogenesis',Description)) 

gsea_plot <- gsea %>%
  ggbarplot(x = 'Description', y = 'neg_log_q',
            xlab = '', ylab = '-log10(q)',
            color = 'white', fill = '#00468B') +
  coord_flip() +
  theme(text = element_text(family = our_font),  
        axis.text.x=element_text(size=axis_text_size),
        axis.title.x=element_blank(),
        axis.text.y = element_text(size = 8),
        legend.position="none") 

```

## Assemble

```{r out.width=800}
g <- plot_grid(compare_plot, pos, len, gc, strength, gsea_plot,  
          nrow = 3, 
          align = 'hv',
          labels = 'AUTO',
          axis = 'b')  +
  ggsave('../figs/d5_Assembled_fig4_intersection.jpg', width = 8, height = 12, dpi = 300)
```
